Improvements to dump_db.py
- sort the morphs in the output. The output is first sorted by morph,
  and within morph, by noteid

- optionally output the locations.

- rename script from dumb_py to dump_db.py (what "dump" typo)

- update docs/new-features.org accordingly
