+++ docs/new-features.org
@@ -57,35 +57,37 @@ the script scripts/dump_db.py can be used to dump the pickled databases (.db)
 from the command line. Simply run 
 
 #+begin_src bash
-python3 scripts/dump_db.py <picked-database-file>
+python3 scripts/dump_db.py <picked-database-file> locations
 #+end_src
 
 The output is straighforward:
 
 For every moprh, it prints its information and the noteid, fieldname, fieldvalue, maturity (in days), guid and the weight.
-See the documentation of Location (under sqlite3) above
+See the documentation of Location (under sqlite3) above.
+
+Output is sorted by morph, the locations. If the locations parameter is not specified, only output morphs
 
 #+begin_example
 norm	base	inflected	reading	pos	subpos
 が	が	が	ガ	助詞	格助詞
-	 	1536650898064;sentenceJPnuke;45;DJgfwS>YL^;1
-	 	1536651281457;sentenceJp;22;o!aP/}@6+5;1
-	 	1536650897940;sentenceJp;259;d>&JX6S_Bd;1
-	 	1487630686913;Expression;277;v-,8t@ro!L;1
-	 	1487630686856;Expression;168;mX1IuArU9S;1
-	 	1489462758391;Expression;103;fd[asSdN3~;1
-	 	1487630687241;Expression;320;gTy?mxI5y];1
-	 	1536651300085;sentenceJp;22;C.!^yST7m8;1
-	 	1536651299859;sentenceJp;200;CcH+l<Y_td;1
-	 	1487630686818;Expression;138;q^h7dzq%1n;1
-	 	1489462776809;Expression;114;g@A7}LZ#;&;1
-	 	1536651300164;sentenceJp;22;fnM.#k3}En;1
-	 	1536650897908;sentenceJp;147;i[{pI;:U>|;1
-	 	1536650898337;sentenceJp;23;ez^L$2ON)M;1
-	 	1489462758327;Expression;198;lQ@.k1)U[8;1
-	 	1487630687083;Expression;288;nD){h(39!I;1
-	 	1489462791313;Expression;54;IR)(c(9dd6;1
-	 	1487630686948;Expression;193;kux^L#r]R4;1
+	 	1487630686818	Expression	138	q^h7dzq%1n	1
+	 	1487630686856	Expression	168	mX1IuArU9S	1
+	 	1487630686913	Expression	277	v-,8t@ro!L	1
+	 	1487630686948	Expression	193	kux^L#r]R4	1
+	 	1487630687083	Expression	288	nD){h(39!I	1
+	 	1487630687241	Expression	320	gTy?mxI5y]	1
+	 	1489462758327	Expression	198	lQ@.k1)U[8	1
+	 	1489462758391	Expression	103	fd[asSdN3~	1
+	 	1489462776809	Expression	114	g@A7}LZ#;&	1
+	 	1489462791313	Expression	54	IR)(c(9dd6	1
+	 	1536650897908	sentenceJp	147	i[{pI;:U>|	1
+	 	1536650897940	sentenceJp	259	d>&JX6S_Bd	1
+	 	1536650898064	sentenceJPnuke	45	DJgfwS>YL^	1
+	 	1536650898337	sentenceJp	23	ez^L$2ON)M	1
+	 	1536651281457	sentenceJp	22	o!aP/}@6+5	1
+	 	1536651299859	sentenceJp	200	CcH+l<Y_td	1
+	 	1536651300085	sentenceJp	22	C.!^yST7m8	1
+	 	1536651300164	sentenceJp	22	fnM.#k3}En	1
 #+end_example
